Refactored user context and added a fetch for clients
